Custom exceptions, no longer thrown if given invalid code value.
Now you have to check whether code value is valid manually if you want
to (using Code#valid?).

Probably the nice behaviour would be to accept any value when reading an
ONIX file, but throw an exception on invalid values when writing out.
But how to detect? One option might be to override to_xml for code
accessors to add the check... hmm.