Update lib/onix/audience_range.rb
Removed the integer setting on audience_range_values as it now allows for number or string entries as per ONIX documentation.

The definition of the <AudienceRangeValue> element has been extended to allowstrings to be carried instead of integers only, so that audience ranges can bespecified in both numeric and non-numeric terms. This is particularly applicable toUS school and pre-school levels.